refactor: address PR review feedback and consolidate tests
- Move inline tests from headless.rs to separate headless_tests.rs per
  CLAUDE.md convention (test file placement in separate files)
- Consolidate 34 tests → 24 by merging duplicative roundtrip and default
  tests into comprehensive single tests with same or better coverage
- Update try_wait/pid doc comments to document Codex behavior
- Fix stderr_cursor initialization to use current tail (prevents stale
  replay for new sessions)
- Replace OnceCell with Mutex<Option> for CODEX_RUNTIME to allow
  detecting dead runtime and re-spawning
- Add Codex-aware channel-alive check in reconcile_process_health
